move directory, rename cff
