+++ CatProducer/python/physicsObjectRecipes/egmSmearing_cff.py
