+++ CatProducer/python/physicsObjectRecipes/egmVersionedID_cff.py
